:recycle: refactor(User): - User Service Command/Query 분리 - Quiz, Coin Service에 User 누적 정보 로직 추가 - User Update 로직 수정 (#4)
